requirements.txt
fastapi>=0.118.3
uvicorn[standard]>=0.35.0
jinja2>=3.1.6
python-multipart>=0.0.20
sqlalchemy>=2.0.40
itsdangerous>=2.2.0
email-validator>=2.2.0
bcrypt>=5.0.0
pwdlib[argon2]>=0.3.0
python-dotenv>=1.1.0
psycopg[binary]>=3.2.9
tzdata>=2025.2
aiogram>=3.18.0
httpx>=0.28.1
